requirements.txt
fastapi
uvicorn
python-multipart
pdfplumber
python-docx
groq
spacy
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl
jinja2
python-dotenv